backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
joblib==1.4.2
xgboost==2.1.4
gunicorn==23.0.0

